change sendPresence from sendMessage to chat
