+++ .vscode/settings.json
@@ -5,8 +5,8 @@
   "editor.smoothScrolling": true,
   "editor.tabSize": 2,
   "editor.codeActionsOnSave": {
-     "source.fixAll.eslint": true,
-     "source.fixAll": true
+    "source.fixAll.eslint": "explicit",
+    "source.fixAll": "explicit"
   },
   "prisma-smart-formatter.typescript.defaultFormatter": "esbenp.prettier-vscode",
   "prisma-smart-formatter.prisma.defaultFormatter": "Prisma.prisma"
