+++ src/whatsapp/dto/sendMessage.dto.ts
@@ -46,12 +46,7 @@ class PollMessage {
   values: string[];
   messageSecret?: Uint8Array;
 }
-export class SendPresenceDto extends Metadata {
-  options: {
-    presence: WAPresence;
-    delay: number;
-  };
-}
+
 export class SendTextDto extends Metadata {
   textMessage: TextMessage;
 }
